Componentize HasChainStatusType
