+++ Cargo.toml
@@ -6,12 +6,12 @@ members = [
     "crates/relayer-components",
     "crates/relayer-components-extra",
     "crates/relayer-subscription",
-    "crates/relayer-mock",
     "crates/relayer-runtime",
     "crates/cosmos-client-components",
     "crates/relayer-cosmos",
     "crates/relayer-solomachine",
     "crates/relayer-cosmos-mock",
+    "crates/relayer-mock",
     "tools/integration-test",
     "tools/test-framework",
 ]
