+++ crates/relayer-components/src/chain/traits/types/status.rs
@@ -1,6 +1,7 @@
 use cgp_core::prelude::*;
 
-use crate::chain::traits::types::chain::HasChainTypes;
+use crate::chain::traits::types::height::HasHeightType;
+use crate::chain::traits::types::timestamp::HasTimestampType;
 
 /**
    A chain context that offers a [`ChainStatus`](Self::ChainStatus) type to
@@ -25,8 +26,8 @@ use crate::chain::traits::types::chain::HasChainTypes;
    chain status queries can be cached without needing to know what information
    is contained inside the chain status.
 */
-// #[derive_component(ChainStatusTypeProviderComponent, ChainStatusTypeProvider<Chain>)]
-pub trait HasChainStatusType: HasChainTypes {
+#[derive_component(ChainStatusTypeProviderComponent, ChainStatusTypeProvider<Chain>)]
+pub trait HasChainStatusType: HasHeightType + HasTimestampType {
     /**
        Contains information about the current status of the blockchain.
     */
