+++ crates/relayer-cosmos/src/impls/chain/component.rs
@@ -57,6 +57,7 @@ use ibc_relayer_components::chain::traits::components::update_client_payload_bui
 use ibc_relayer_components::chain::traits::components::write_ack_querier::WriteAckQuerierComponent;
 use ibc_relayer_components::chain::traits::types::height::HeightTypeProviderComponent;
 use ibc_relayer_components::chain::traits::types::message::MessageTypeProviderComponent;
+use ibc_relayer_components::chain::traits::types::status::ChainStatusTypeProviderComponent;
 use ibc_relayer_components_extra::components::extra::chain::ExtraChainComponents;
 use ibc_relayer_components_extra::components::extra::closures::chain::all::CanUseExtraChainComponents;
 
@@ -88,6 +89,7 @@ delegate_components!(
     [
         HeightTypeProviderComponent,
         MessageTypeProviderComponent,
+        ChainStatusTypeProviderComponent,
     ]:
         ProvideCosmosChainTypes,
     MessageSenderComponent:
